support hydrate, ssr apps
