+++ Cargo.toml
@@ -11,10 +11,17 @@ keywords=["leptos", "hotkeys", "wasm"]
 
 
 [dependencies]
-leptos = { version = "0.6.5", features = ["csr", "nightly"] }
-wasm-bindgen = "0.2.89"
+leptos = { version = "0.6.5", features = ["nightly"] }
+wasm-bindgen = { version = "0.2.89", optional = true }
 log = "0.4.20"
+cfg-if = { version = "1.0.0", features = [] }
+web-sys = { version = "0.3.66", optional = true }
 
 [lib]
 name = "leptos_hotkeys"
 path = "src/lib.rs"
+
+[features]
+hydrate = ["dep:web-sys", "dep:wasm-bindgen"]
+csr = ["dep:web-sys", "dep:wasm-bindgen"]
+ssr = []
