--- example/demo/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "demo"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-leptos = { version = "0.6.5", features = ["csr", "nightly"] }
-console_error_panic_hook = "0.1.7"
-console_log = "1.0.0"
-copy_dir = "0.1.3"
-leptos_meta = "0.6.5"
-leptos_router = { version = "0.6.5", features = ["csr"] }
-log = "0.4.20"
-leptos_theme = "0.1.2"
-leptos_hotkeys = { path = "../.." }
