--- example/demo/index.html
@@ -1,64 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>demo</title>
-    <link data-trunk rel="rust" data-wasm-opt="z" />
-
-    <link data-trunk rel="copy-file" href="vercel.json" />
-    <link data-trunk rel="css" href="/style/output.css" />
-
-    <link rel="preconnect" href="https://fonts.googleapis.com">
-    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-    <link href="https://fonts.googleapis.com/css2?family=Roboto+Mono&display=swap" rel="stylesheet">
-
-    <style>
-        .green, .yellow {
-            padding: 12px;
-            cursor: pointer;
-        }
-
-        .green {
-            background-color: #39ff14;
-        }
-
-        .yellow {
-            background-color: yellow;
-        }
-
-        .green:focus, .yellow:focus {
-            outline: none;
-            box-shadow: 0 0 4px rgba(0, 0, 0, 0.6);
-        }
-
-
-        @media (prefers-color-scheme: light) {
-            .not-active {
-                background-color: #fdfdfd;
-                color: #1a1a1a;
-            }
-
-            .active {
-                background-color: #1a1a1a;
-                color: #fdfdfd
-            }
-        }
-
-        @media (prefers-color-scheme: dark) {
-            .active {
-                background-color: #fdfdfd;
-                color: #1a1a1a;
-            }
-
-            .not-active {
-                background-color: #1a1a1a;
-                color: #fdfdfd
-            }
-        }
-    </style>
-        </head>
-<body>
-</body>
-
-</html>
