--- example/demo/rust-toolchain.toml
@@ -1,3 +1,0 @@
-
-[toolchain]
-channel = "nightly"
