--- example/demo/src/main.rs
@@ -1,14 +1,0 @@
-use app::*;
-use leptos::*;
-mod app;
-
-pub fn main() {
-    _ = console_log::init_with_level(log::Level::Debug);
-    console_error_panic_hook::set_once();
-
-    logging::log!("csr mode - mounting to body");
-
-    mount_to_body(|| {
-        view! { <App/> }
-    });
-}
