+++ src/hotkeys_provider.rs
@@ -1,23 +1,41 @@
 use crate::scopes;
-
+use cfg_if::cfg_if;
 use leptos::html::div;
-use leptos::web_sys::KeyboardEvent;
 use leptos::*;
 use std::collections::{HashMap, HashSet};
-use wasm_bindgen::closure::Closure;
-use wasm_bindgen::JsCast;
-use web_sys::EventTarget;
+
+cfg_if! {
+    if #[cfg(any(feature = "hydrate", feature= "csr"))] {
+        use wasm_bindgen::closure::Closure;
+        use wasm_bindgen::JsCast;
+        use web_sys::{
+            EventTarget, KeyboardEvent
+        };
+    }
+}
 
 // Defining a hotkey context structure
-#[derive(Clone)]
-pub struct HotkeysContext {
-    pub(crate) pressed_keys: RwSignal<HashMap<String, KeyboardEvent>>,
-    pub active_ref_target: RwSignal<Option<EventTarget>>,
-    pub set_ref_target: Callback<Option<EventTarget>>,
-    pub active_scopes: RwSignal<HashSet<String>>,
-    pub enable_scope: Callback<String>,
-    pub disable_scope: Callback<String>,
-    pub toggle_scope: Callback<String>,
+cfg_if! {
+    if #[cfg(any(feature = "hydrate", feature= "csr"))] {
+        #[derive(Clone)]
+        pub struct HotkeysContext {
+            pub(crate) pressed_keys: RwSignal<HashMap<String, KeyboardEvent>>,
+            pub active_ref_target: RwSignal<Option<EventTarget>>,
+            pub set_ref_target: Callback<Option<EventTarget>>,
+            pub active_scopes: RwSignal<HashSet<String>>,
+            pub enable_scope: Callback<String>,
+            pub disable_scope: Callback<String>,
+            pub toggle_scope: Callback<String>,
+        }
+    } else {
+        #[derive(Clone)]
+        pub struct HotkeysContext {
+            pub active_scopes: RwSignal<HashSet<String>>,
+            pub enable_scope: Callback<String>,
+            pub disable_scope: Callback<String>,
+            pub toggle_scope: Callback<String>,
+        }
+    }
 }
 
 pub fn use_hotkeys_context() -> HotkeysContext {
