+++ src/macros.rs
@@ -37,7 +37,7 @@ macro_rules! use_hotkeys {
     (($key_combo:literal) => $($code:tt)*) => {
         {
             use_hotkeys_scoped(
-                $key_combo.to_string(), 
+                $key_combo.to_string(),
                 Callback::new(
                 $($code)*
                 ),
@@ -49,7 +49,7 @@ macro_rules! use_hotkeys {
     (($key_combo:expr) => $($code:tt)*) => {
         {
             use_hotkeys_scoped(
-                $key_combo.to_string(), 
+                $key_combo.to_string(),
                 Callback::new(
                 $($code)*
                 ),
@@ -61,8 +61,8 @@ macro_rules! use_hotkeys {
     (($key_combo:expr $(, $scopes:literal)*) => $($code:tt)*) => {
         {
             use_hotkeys_scoped(
-                $key_combo.to_string(), 
-                Callback::new( 
+                $key_combo.to_string(),
+                Callback::new(
                     $($code)*
                 ),
                 vec![$($scopes.to_string(),)*]
@@ -73,8 +73,8 @@ macro_rules! use_hotkeys {
     (($key_combo:literal $(, $scopes:literal)*) => $($code:tt)*) => {
         {
             use_hotkeys_scoped(
-                $key_combo.to_string(), 
-                Callback::new( 
+                $key_combo.to_string(),
+                Callback::new(
                     $($code)*
                 ),
                 vec![$($scopes.to_string(),)*]
@@ -85,8 +85,8 @@ macro_rules! use_hotkeys {
     (($key_combo:literal $(, $scopes:expr)*) => $($code:tt)*) => {
         {
             use_hotkeys_scoped(
-                $key_combo.to_string(), 
-                Callback::new( 
+                $key_combo.to_string(),
+                Callback::new(
                     $($code)*
                 ),
                 vec![$($scopes.to_string(),)*]
@@ -97,8 +97,8 @@ macro_rules! use_hotkeys {
     (($key_combo:expr $(, $scopes:expr)*) => $($code:tt)*) => {
         {
             use_hotkeys_scoped(
-                $key_combo.to_string(), 
-                Callback::new( 
+                $key_combo.to_string(),
+                Callback::new(
                     $($code)*
                 ),
                 vec![$($scopes.to_string(),)*]
@@ -114,7 +114,7 @@ macro_rules! use_hotkeys_ref {
     (($key_combo:literal) => $($code:tt)*) => {
         {
             use_hotkeys_ref_scoped(
-                $key_combo.to_string(), 
+                $key_combo.to_string(),
                 Callback::new(
                 $($code)*
                 ),
@@ -126,7 +126,7 @@ macro_rules! use_hotkeys_ref {
     (($key_combo:expr) => $($code:tt)*) => {
         {
             use_hotkeys_ref_scoped(
-                $key_combo.to_string(), 
+                $key_combo.to_string(),
                 Callback::new(
                 $($code)*
                 ),
@@ -138,8 +138,8 @@ macro_rules! use_hotkeys_ref {
     (($key_combo:expr $(, $scopes:literal)*) => $($code:tt)*) => {
         {
             use_hotkeys_ref_scoped(
-                $key_combo.to_string(), 
-                Callback::new( 
+                $key_combo.to_string(),
+                Callback::new(
                     $($code)*
                 ),
                 vec![$($scopes.to_string(),)*]
@@ -150,8 +150,8 @@ macro_rules! use_hotkeys_ref {
     (($key_combo:literal $(, $scopes:literal)*) => $($code:tt)*) => {
         {
             use_hotkeys_ref_scoped(
-                $key_combo.to_string(), 
-                Callback::new( 
+                $key_combo.to_string(),
+                Callback::new(
                     $($code)*
                 ),
                 vec![$($scopes.to_string(),)*]
@@ -162,8 +162,8 @@ macro_rules! use_hotkeys_ref {
     (($key_combo:literal $(, $scopes:expr)*) => $($code:tt)*) => {
         {
             use_hotkeys_ref_scoped(
-                $key_combo.to_string(), 
-                Callback::new( 
+                $key_combo.to_string(),
+                Callback::new(
                     $($code)*
                 ),
                 vec![$($scopes.to_string(),)*]
@@ -174,8 +174,8 @@ macro_rules! use_hotkeys_ref {
     (($key_combo:expr $(, $scopes:expr)*) => $($code:tt)*) => {
         {
             use_hotkeys_ref_scoped(
-                $key_combo.to_string(), 
-                Callback::new( 
+                $key_combo.to_string(),
+                Callback::new(
                     $($code)*
                 ),
                 vec![$($scopes.to_string(),)*]
