+++ workspace/ssr-demo/style/main.scss
@@ -1,0 +1,4 @@
+body {
+	font-family: sans-serif;
+	text-align: center;
+}
